requirements.txt
Flask
Gunicorn
psycopg2-binary
pandas
uuid
statistics